+++ src/Dimensions/LookupDimension.php
@@ -9,12 +9,17 @@ class LookupDimension implements DimensionInterface
 
     protected string $outputName;
 
-    protected string $registeredLookupFunction;
+    /**
+     * @var string|array<int|string,string>
+     */
+    protected $registeredLookupFunctionOrMap;
 
     /**
      * @var bool|string
      */
     protected $keepMissingValue;
+
+    protected bool $isOneToOne;
 
     /**
      * DefaultDimension constructor.
@@ -26,27 +31,29 @@ class LookupDimension implements DimensionInterface
      * it to a string it wil replace the missing value with that string. It defaults to false
      * which will ignore the value.
      *
-     * @param string      $dimension
-     * @param string      $registeredLookupFunction
-     * @param string      $outputName
-     * @param bool|string $keepMissingValue
+     * @param string                          $dimension
+     * @param string|array<int|string,string> $registeredLookupFunctionOrMap
+     * @param string                          $outputName
+     * @param bool|string                     $keepMissingValue
      */
     public function __construct(
         string $dimension,
-        string $registeredLookupFunction,
+        $registeredLookupFunctionOrMap,
         string $outputName = '',
-        $keepMissingValue = false
+        $keepMissingValue = false,
+        bool $isOneToOne = false
     ) {
-        $this->dimension                = $dimension;
-        $this->outputName               = $outputName ?: $dimension;
-        $this->registeredLookupFunction = $registeredLookupFunction;
-        $this->keepMissingValue         = $keepMissingValue;
+        $this->dimension                     = $dimension;
+        $this->outputName                    = $outputName ?: $dimension;
+        $this->registeredLookupFunctionOrMap = $registeredLookupFunctionOrMap;
+        $this->keepMissingValue              = $keepMissingValue;
+        $this->isOneToOne                    = $isOneToOne;
     }
 
     /**
      * Return the dimension as it should be used in a druid query.
      *
-     * @return array<string,string|bool>
+     * @return array<string,string|bool|array<string,string|bool|string[]>>
      */
     public function toArray(): array
     {
@@ -54,8 +61,17 @@ class LookupDimension implements DimensionInterface
             'type'       => 'lookup',
             'dimension'  => $this->dimension,
             'outputName' => $this->outputName,
-            'name'       => $this->registeredLookupFunction,
         ];
+
+        if (is_string($this->registeredLookupFunctionOrMap)) {
+            $result['name'] = $this->registeredLookupFunctionOrMap;
+        } elseif (is_array($this->registeredLookupFunctionOrMap)) {
+            $result['lookup'] = [
+                'type'       => 'map',
+                'map'        => $this->registeredLookupFunctionOrMap,
+                'isOneToOne' => $this->isOneToOne,
+            ];
+        }
 
         if ($this->keepMissingValue === true) {
             $result['retainMissingValue'] = true;